initial preparation refactor
