+++ crates/bevy_render/src/lib.rs
@@ -29,7 +29,6 @@ pub mod gpu_readback;
 pub mod mesh;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod pipelined_rendering;
-pub mod primitives;
 pub mod render_asset;
 pub mod render_graph;
 pub mod render_phase;
@@ -66,6 +65,7 @@ pub mod prelude {
     };
 }
 
+pub use crate::mesh::bounding as primitives;
 use batching::gpu_preprocessing::BatchingPlugin;
 use bevy_ecs::schedule::ScheduleBuildSettings;
 use bevy_utils::prelude::default;
