+++ crates/bevy_render/src/mesh/conversions.rs
@@ -24,7 +24,7 @@
 //! assert!(error.is_err());
 //! ```
 
-use crate::mesh::VertexAttributeValues;
+use super::VertexAttributeValues;
 use bevy_math::{IVec2, IVec3, IVec4, UVec2, UVec3, UVec4, Vec2, Vec3, Vec3A, Vec4};
 use thiserror::Error;
