+++ crates/bevy_render/src/mesh/mesh.rs
@@ -1,32 +1,21 @@
-mod conversions;
-pub mod skinning;
 use bevy_transform::components::Transform;
-use bitflags::bitflags;
 pub use wgpu::PrimitiveTopology;
 
-use crate::{
-    prelude::Image,
-    primitives::Aabb,
-    render_asset::{PrepareAssetError, RenderAsset, RenderAssetUsages, RenderAssets},
-    render_resource::{TextureView, VertexBufferLayout},
-    texture::GpuImage,
+use super::{
+    bounding::Aabb, face_normal, generate_tangents_for_mesh, scale_normal, FourIterators,
+    GenerateTangentsError, Indices, MeshAttributeData, MeshTrianglesError, MeshVertexAttribute,
+    MeshVertexAttributeId, MeshVertexBufferLayout, MeshVertexBufferLayoutRef,
+    MeshVertexBufferLayouts, MeshWindingInvertError, VertexAttributeValues, VertexBufferLayout,
+    VertexFormatSize,
 };
 use alloc::collections::BTreeMap;
-use bevy_asset::{Asset, Handle};
-use bevy_derive::EnumVariantMeta;
-use bevy_ecs::system::{
-    lifetimeless::{SRes, SResMut},
-    SystemParamItem,
-};
+use bevy_asset::{Asset, Handle, RenderAssetUsages};
+use bevy_image::Image;
 use bevy_math::{primitives::Triangle3d, *};
 use bevy_reflect::Reflect;
-use bevy_utils::tracing::{error, warn};
+use bevy_utils::tracing::warn;
 use bytemuck::cast_slice;
-use core::{hash::Hash, iter::FusedIterator};
-use thiserror::Error;
-use wgpu::{IndexFormat, VertexAttribute, VertexFormat, VertexStepMode};
-
-use super::{MeshVertexBufferLayoutRef, MeshVertexBufferLayouts};
+use wgpu::{VertexAttribute, VertexFormat, VertexStepMode};
 
 pub const INDEX_BUFFER_ASSET_INDEX: u64 = 0;
 pub const VERTEX_ATTRIBUTE_BUFFER_ID: u64 = 10;
@@ -1087,6 +1076,10 @@ impl Mesh {
         self.morph_targets = Some(morph_targets);
     }
 
+    pub fn morph_targets(&self) -> Option<&Handle<Image>> {
+        self.morph_targets.as_ref()
+    }
+
     /// Consumes the mesh and returns a mesh with the given [morph targets].
     ///
     /// This requires a "morph target image". See [`MorphTargetImage`](crate::mesh::morph::MorphTargetImage) for info.
@@ -1226,763 +1219,11 @@ impl Mesh {
     }
 }
 
-/// A disjunction of four iterators. This is necessary to have a well-formed type for the output
-/// of [`Mesh::triangles`], which produces iterators of four different types depending on the
-/// branch taken.
-enum FourIterators<A, B, C, D> {
-    First(A),
-    Second(B),
-    Third(C),
-    Fourth(D),
-}
-
-impl<A, B, C, D, I> Iterator for FourIterators<A, B, C, D>
-where
-    A: Iterator<Item = I>,
-    B: Iterator<Item = I>,
-    C: Iterator<Item = I>,
-    D: Iterator<Item = I>,
-{
-    type Item = I;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            FourIterators::First(iter) => iter.next(),
-            FourIterators::Second(iter) => iter.next(),
-            FourIterators::Third(iter) => iter.next(),
-            FourIterators::Fourth(iter) => iter.next(),
-        }
-    }
-}
-
-/// An error that occurred while trying to invert the winding of a [`Mesh`].
-#[derive(Debug, Error)]
-pub enum MeshWindingInvertError {
-    /// This error occurs when you try to invert the winding for a mesh with [`PrimitiveTopology::PointList`].
-    #[error("Mesh winding invertation does not work for primitive topology `PointList`")]
-    WrongTopology,
-
-    /// This error occurs when you try to invert the winding for a mesh with
-    /// * [`PrimitiveTopology::TriangleList`], but the indices are not in chunks of 3.
-    /// * [`PrimitiveTopology::LineList`], but the indices are not in chunks of 2.
-    #[error("Indices weren't in chunks according to topology")]
-    AbruptIndicesEnd,
-}
-
-/// An error that occurred while trying to extract a collection of triangles from a [`Mesh`].
-#[derive(Debug, Error)]
-pub enum MeshTrianglesError {
-    #[error("Source mesh does not have primitive topology TriangleList or TriangleStrip")]
-    WrongTopology,
-
-    #[error("Source mesh lacks position data")]
-    MissingPositions,
-
-    #[error("Source mesh position data is not Float32x3")]
-    PositionsFormat,
-
-    #[error("Source mesh lacks face index data")]
-    MissingIndices,
-
-    #[error("Face index data references vertices that do not exist")]
-    BadIndices,
-}
-
 impl core::ops::Mul<Mesh> for Transform {
     type Output = Mesh;
 
     fn mul(self, rhs: Mesh) -> Self::Output {
         rhs.transformed_by(self)
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct MeshVertexAttribute {
-    /// The friendly name of the vertex attribute
-    pub name: &'static str,
-
-    /// The _unique_ id of the vertex attribute. This will also determine sort ordering
-    /// when generating vertex buffers. Built-in / standard attributes will use "close to zero"
-    /// indices. When in doubt, use a random / very large u64 to avoid conflicts.
-    pub id: MeshVertexAttributeId,
-
-    /// The format of the vertex attribute.
-    pub format: VertexFormat,
-}
-
-impl MeshVertexAttribute {
-    pub const fn new(name: &'static str, id: u64, format: VertexFormat) -> Self {
-        Self {
-            name,
-            id: MeshVertexAttributeId(id),
-            format,
-        }
-    }
-
-    pub const fn at_shader_location(&self, shader_location: u32) -> VertexAttributeDescriptor {
-        VertexAttributeDescriptor::new(shader_location, self.id, self.name)
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Ord, PartialOrd, Hash)]
-pub struct MeshVertexAttributeId(u64);
-
-impl From<MeshVertexAttribute> for MeshVertexAttributeId {
-    fn from(attribute: MeshVertexAttribute) -> Self {
-        attribute.id
-    }
-}
-
-#[derive(Debug, Clone, Hash, Eq, PartialEq)]
-pub struct MeshVertexBufferLayout {
-    attribute_ids: Vec<MeshVertexAttributeId>,
-    layout: VertexBufferLayout,
-}
-
-impl MeshVertexBufferLayout {
-    pub fn new(attribute_ids: Vec<MeshVertexAttributeId>, layout: VertexBufferLayout) -> Self {
-        Self {
-            attribute_ids,
-            layout,
-        }
-    }
-
-    #[inline]
-    pub fn contains(&self, attribute_id: impl Into<MeshVertexAttributeId>) -> bool {
-        self.attribute_ids.contains(&attribute_id.into())
-    }
-
-    #[inline]
-    pub fn attribute_ids(&self) -> &[MeshVertexAttributeId] {
-        &self.attribute_ids
-    }
-
-    #[inline]
-    pub fn layout(&self) -> &VertexBufferLayout {
-        &self.layout
-    }
-
-    pub fn get_layout(
-        &self,
-        attribute_descriptors: &[VertexAttributeDescriptor],
-    ) -> Result<VertexBufferLayout, MissingVertexAttributeError> {
-        let mut attributes = Vec::with_capacity(attribute_descriptors.len());
-        for attribute_descriptor in attribute_descriptors {
-            if let Some(index) = self
-                .attribute_ids
-                .iter()
-                .position(|id| *id == attribute_descriptor.id)
-            {
-                let layout_attribute = &self.layout.attributes[index];
-                attributes.push(VertexAttribute {
-                    format: layout_attribute.format,
-                    offset: layout_attribute.offset,
-                    shader_location: attribute_descriptor.shader_location,
-                });
-            } else {
-                return Err(MissingVertexAttributeError {
-                    id: attribute_descriptor.id,
-                    name: attribute_descriptor.name,
-                    pipeline_type: None,
-                });
-            }
-        }
-
-        Ok(VertexBufferLayout {
-            array_stride: self.layout.array_stride,
-            step_mode: self.layout.step_mode,
-            attributes,
-        })
-    }
-}
-
-#[derive(Error, Debug)]
-#[error("Mesh is missing requested attribute: {name} ({id:?}, pipeline type: {pipeline_type:?})")]
-pub struct MissingVertexAttributeError {
-    pub(crate) pipeline_type: Option<&'static str>,
-    id: MeshVertexAttributeId,
-    name: &'static str,
-}
-
-pub struct VertexAttributeDescriptor {
-    pub shader_location: u32,
-    pub id: MeshVertexAttributeId,
-    name: &'static str,
-}
-
-impl VertexAttributeDescriptor {
-    pub const fn new(shader_location: u32, id: MeshVertexAttributeId, name: &'static str) -> Self {
-        Self {
-            shader_location,
-            id,
-            name,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct MeshAttributeData {
-    attribute: MeshVertexAttribute,
-    values: VertexAttributeValues,
-}
-
-fn face_normal(a: [f32; 3], b: [f32; 3], c: [f32; 3]) -> [f32; 3] {
-    let (a, b, c) = (Vec3::from(a), Vec3::from(b), Vec3::from(c));
-    (b - a).cross(c - a).normalize().into()
-}
-
-pub trait VertexFormatSize {
-    fn get_size(self) -> u64;
-}
-
-impl VertexFormatSize for VertexFormat {
-    #[allow(clippy::match_same_arms)]
-    fn get_size(self) -> u64 {
-        match self {
-            VertexFormat::Uint8x2 => 2,
-            VertexFormat::Uint8x4 => 4,
-            VertexFormat::Sint8x2 => 2,
-            VertexFormat::Sint8x4 => 4,
-            VertexFormat::Unorm8x2 => 2,
-            VertexFormat::Unorm8x4 => 4,
-            VertexFormat::Snorm8x2 => 2,
-            VertexFormat::Snorm8x4 => 4,
-            VertexFormat::Unorm10_10_10_2 => 4,
-            VertexFormat::Uint16x2 => 2 * 2,
-            VertexFormat::Uint16x4 => 2 * 4,
-            VertexFormat::Sint16x2 => 2 * 2,
-            VertexFormat::Sint16x4 => 2 * 4,
-            VertexFormat::Unorm16x2 => 2 * 2,
-            VertexFormat::Unorm16x4 => 2 * 4,
-            VertexFormat::Snorm16x2 => 2 * 2,
-            VertexFormat::Snorm16x4 => 2 * 4,
-            VertexFormat::Float16x2 => 2 * 2,
-            VertexFormat::Float16x4 => 2 * 4,
-            VertexFormat::Float32 => 4,
-            VertexFormat::Float32x2 => 4 * 2,
-            VertexFormat::Float32x3 => 4 * 3,
-            VertexFormat::Float32x4 => 4 * 4,
-            VertexFormat::Uint32 => 4,
-            VertexFormat::Uint32x2 => 4 * 2,
-            VertexFormat::Uint32x3 => 4 * 3,
-            VertexFormat::Uint32x4 => 4 * 4,
-            VertexFormat::Sint32 => 4,
-            VertexFormat::Sint32x2 => 4 * 2,
-            VertexFormat::Sint32x3 => 4 * 3,
-            VertexFormat::Sint32x4 => 4 * 4,
-            VertexFormat::Float64 => 8,
-            VertexFormat::Float64x2 => 8 * 2,
-            VertexFormat::Float64x3 => 8 * 3,
-            VertexFormat::Float64x4 => 8 * 4,
-        }
-    }
-}
-
-/// Contains an array where each entry describes a property of a single vertex.
-/// Matches the [`VertexFormats`](VertexFormat).
-#[derive(Clone, Debug, EnumVariantMeta)]
-pub enum VertexAttributeValues {
-    Float32(Vec<f32>),
-    Sint32(Vec<i32>),
-    Uint32(Vec<u32>),
-    Float32x2(Vec<[f32; 2]>),
-    Sint32x2(Vec<[i32; 2]>),
-    Uint32x2(Vec<[u32; 2]>),
-    Float32x3(Vec<[f32; 3]>),
-    Sint32x3(Vec<[i32; 3]>),
-    Uint32x3(Vec<[u32; 3]>),
-    Float32x4(Vec<[f32; 4]>),
-    Sint32x4(Vec<[i32; 4]>),
-    Uint32x4(Vec<[u32; 4]>),
-    Sint16x2(Vec<[i16; 2]>),
-    Snorm16x2(Vec<[i16; 2]>),
-    Uint16x2(Vec<[u16; 2]>),
-    Unorm16x2(Vec<[u16; 2]>),
-    Sint16x4(Vec<[i16; 4]>),
-    Snorm16x4(Vec<[i16; 4]>),
-    Uint16x4(Vec<[u16; 4]>),
-    Unorm16x4(Vec<[u16; 4]>),
-    Sint8x2(Vec<[i8; 2]>),
-    Snorm8x2(Vec<[i8; 2]>),
-    Uint8x2(Vec<[u8; 2]>),
-    Unorm8x2(Vec<[u8; 2]>),
-    Sint8x4(Vec<[i8; 4]>),
-    Snorm8x4(Vec<[i8; 4]>),
-    Uint8x4(Vec<[u8; 4]>),
-    Unorm8x4(Vec<[u8; 4]>),
-}
-
-impl VertexAttributeValues {
-    /// Returns the number of vertices in this [`VertexAttributeValues`]. For a single
-    /// mesh, all of the [`VertexAttributeValues`] must have the same length.
-    #[allow(clippy::match_same_arms)]
-    pub fn len(&self) -> usize {
-        match self {
-            VertexAttributeValues::Float32(values) => values.len(),
-            VertexAttributeValues::Sint32(values) => values.len(),
-            VertexAttributeValues::Uint32(values) => values.len(),
-            VertexAttributeValues::Float32x2(values) => values.len(),
-            VertexAttributeValues::Sint32x2(values) => values.len(),
-            VertexAttributeValues::Uint32x2(values) => values.len(),
-            VertexAttributeValues::Float32x3(values) => values.len(),
-            VertexAttributeValues::Sint32x3(values) => values.len(),
-            VertexAttributeValues::Uint32x3(values) => values.len(),
-            VertexAttributeValues::Float32x4(values) => values.len(),
-            VertexAttributeValues::Sint32x4(values) => values.len(),
-            VertexAttributeValues::Uint32x4(values) => values.len(),
-            VertexAttributeValues::Sint16x2(values) => values.len(),
-            VertexAttributeValues::Snorm16x2(values) => values.len(),
-            VertexAttributeValues::Uint16x2(values) => values.len(),
-            VertexAttributeValues::Unorm16x2(values) => values.len(),
-            VertexAttributeValues::Sint16x4(values) => values.len(),
-            VertexAttributeValues::Snorm16x4(values) => values.len(),
-            VertexAttributeValues::Uint16x4(values) => values.len(),
-            VertexAttributeValues::Unorm16x4(values) => values.len(),
-            VertexAttributeValues::Sint8x2(values) => values.len(),
-            VertexAttributeValues::Snorm8x2(values) => values.len(),
-            VertexAttributeValues::Uint8x2(values) => values.len(),
-            VertexAttributeValues::Unorm8x2(values) => values.len(),
-            VertexAttributeValues::Sint8x4(values) => values.len(),
-            VertexAttributeValues::Snorm8x4(values) => values.len(),
-            VertexAttributeValues::Uint8x4(values) => values.len(),
-            VertexAttributeValues::Unorm8x4(values) => values.len(),
-        }
-    }
-
-    /// Returns `true` if there are no vertices in this [`VertexAttributeValues`].
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Returns the values as float triples if possible.
-    pub fn as_float3(&self) -> Option<&[[f32; 3]]> {
-        match self {
-            VertexAttributeValues::Float32x3(values) => Some(values),
-            _ => None,
-        }
-    }
-
-    // TODO: add vertex format as parameter here and perform type conversions
-    /// Flattens the [`VertexAttributeValues`] into a sequence of bytes. This is
-    /// useful for serialization and sending to the GPU.
-    #[allow(clippy::match_same_arms)]
-    pub fn get_bytes(&self) -> &[u8] {
-        match self {
-            VertexAttributeValues::Float32(values) => cast_slice(values),
-            VertexAttributeValues::Sint32(values) => cast_slice(values),
-            VertexAttributeValues::Uint32(values) => cast_slice(values),
-            VertexAttributeValues::Float32x2(values) => cast_slice(values),
-            VertexAttributeValues::Sint32x2(values) => cast_slice(values),
-            VertexAttributeValues::Uint32x2(values) => cast_slice(values),
-            VertexAttributeValues::Float32x3(values) => cast_slice(values),
-            VertexAttributeValues::Sint32x3(values) => cast_slice(values),
-            VertexAttributeValues::Uint32x3(values) => cast_slice(values),
-            VertexAttributeValues::Float32x4(values) => cast_slice(values),
-            VertexAttributeValues::Sint32x4(values) => cast_slice(values),
-            VertexAttributeValues::Uint32x4(values) => cast_slice(values),
-            VertexAttributeValues::Sint16x2(values) => cast_slice(values),
-            VertexAttributeValues::Snorm16x2(values) => cast_slice(values),
-            VertexAttributeValues::Uint16x2(values) => cast_slice(values),
-            VertexAttributeValues::Unorm16x2(values) => cast_slice(values),
-            VertexAttributeValues::Sint16x4(values) => cast_slice(values),
-            VertexAttributeValues::Snorm16x4(values) => cast_slice(values),
-            VertexAttributeValues::Uint16x4(values) => cast_slice(values),
-            VertexAttributeValues::Unorm16x4(values) => cast_slice(values),
-            VertexAttributeValues::Sint8x2(values) => cast_slice(values),
-            VertexAttributeValues::Snorm8x2(values) => cast_slice(values),
-            VertexAttributeValues::Uint8x2(values) => cast_slice(values),
-            VertexAttributeValues::Unorm8x2(values) => cast_slice(values),
-            VertexAttributeValues::Sint8x4(values) => cast_slice(values),
-            VertexAttributeValues::Snorm8x4(values) => cast_slice(values),
-            VertexAttributeValues::Uint8x4(values) => cast_slice(values),
-            VertexAttributeValues::Unorm8x4(values) => cast_slice(values),
-        }
-    }
-}
-
-impl From<&VertexAttributeValues> for VertexFormat {
-    fn from(values: &VertexAttributeValues) -> Self {
-        match values {
-            VertexAttributeValues::Float32(_) => VertexFormat::Float32,
-            VertexAttributeValues::Sint32(_) => VertexFormat::Sint32,
-            VertexAttributeValues::Uint32(_) => VertexFormat::Uint32,
-            VertexAttributeValues::Float32x2(_) => VertexFormat::Float32x2,
-            VertexAttributeValues::Sint32x2(_) => VertexFormat::Sint32x2,
-            VertexAttributeValues::Uint32x2(_) => VertexFormat::Uint32x2,
-            VertexAttributeValues::Float32x3(_) => VertexFormat::Float32x3,
-            VertexAttributeValues::Sint32x3(_) => VertexFormat::Sint32x3,
-            VertexAttributeValues::Uint32x3(_) => VertexFormat::Uint32x3,
-            VertexAttributeValues::Float32x4(_) => VertexFormat::Float32x4,
-            VertexAttributeValues::Sint32x4(_) => VertexFormat::Sint32x4,
-            VertexAttributeValues::Uint32x4(_) => VertexFormat::Uint32x4,
-            VertexAttributeValues::Sint16x2(_) => VertexFormat::Sint16x2,
-            VertexAttributeValues::Snorm16x2(_) => VertexFormat::Snorm16x2,
-            VertexAttributeValues::Uint16x2(_) => VertexFormat::Uint16x2,
-            VertexAttributeValues::Unorm16x2(_) => VertexFormat::Unorm16x2,
-            VertexAttributeValues::Sint16x4(_) => VertexFormat::Sint16x4,
-            VertexAttributeValues::Snorm16x4(_) => VertexFormat::Snorm16x4,
-            VertexAttributeValues::Uint16x4(_) => VertexFormat::Uint16x4,
-            VertexAttributeValues::Unorm16x4(_) => VertexFormat::Unorm16x4,
-            VertexAttributeValues::Sint8x2(_) => VertexFormat::Sint8x2,
-            VertexAttributeValues::Snorm8x2(_) => VertexFormat::Snorm8x2,
-            VertexAttributeValues::Uint8x2(_) => VertexFormat::Uint8x2,
-            VertexAttributeValues::Unorm8x2(_) => VertexFormat::Unorm8x2,
-            VertexAttributeValues::Sint8x4(_) => VertexFormat::Sint8x4,
-            VertexAttributeValues::Snorm8x4(_) => VertexFormat::Snorm8x4,
-            VertexAttributeValues::Uint8x4(_) => VertexFormat::Uint8x4,
-            VertexAttributeValues::Unorm8x4(_) => VertexFormat::Unorm8x4,
-        }
-    }
-}
-/// An array of indices into the [`VertexAttributeValues`] for a mesh.
-///
-/// It describes the order in which the vertex attributes should be joined into faces.
-#[derive(Debug, Clone, Reflect)]
-pub enum Indices {
-    U16(Vec<u16>),
-    U32(Vec<u32>),
-}
-
-impl Indices {
-    /// Returns an iterator over the indices.
-    pub fn iter(&self) -> impl Iterator<Item = usize> + '_ {
-        match self {
-            Indices::U16(vec) => IndicesIter::U16(vec.iter()),
-            Indices::U32(vec) => IndicesIter::U32(vec.iter()),
-        }
-    }
-
-    /// Returns the number of indices.
-    pub fn len(&self) -> usize {
-        match self {
-            Indices::U16(vec) => vec.len(),
-            Indices::U32(vec) => vec.len(),
-        }
-    }
-
-    /// Returns `true` if there are no indices.
-    pub fn is_empty(&self) -> bool {
-        match self {
-            Indices::U16(vec) => vec.is_empty(),
-            Indices::U32(vec) => vec.is_empty(),
-        }
-    }
-}
-
-/// An Iterator for the [`Indices`].
-enum IndicesIter<'a> {
-    U16(core::slice::Iter<'a, u16>),
-    U32(core::slice::Iter<'a, u32>),
-}
-
-impl Iterator for IndicesIter<'_> {
-    type Item = usize;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            IndicesIter::U16(iter) => iter.next().map(|val| *val as usize),
-            IndicesIter::U32(iter) => iter.next().map(|val| *val as usize),
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        match self {
-            IndicesIter::U16(iter) => iter.size_hint(),
-            IndicesIter::U32(iter) => iter.size_hint(),
-        }
-    }
-}
-
-impl<'a> ExactSizeIterator for IndicesIter<'a> {}
-impl<'a> FusedIterator for IndicesIter<'a> {}
-
-impl From<&Indices> for IndexFormat {
-    fn from(indices: &Indices) -> Self {
-        match indices {
-            Indices::U16(_) => IndexFormat::Uint16,
-            Indices::U32(_) => IndexFormat::Uint32,
-        }
-    }
-}
-
-bitflags! {
-    /// Our base mesh pipeline key bits start from the highest bit and go
-    /// downward. The PBR mesh pipeline key bits start from the lowest bit and
-    /// go upward. This allows the PBR bits in the downstream crate `bevy_pbr`
-    /// to coexist in the same field without any shifts.
-    #[derive(Clone, Debug)]
-    pub struct BaseMeshPipelineKey: u64 {
-        const MORPH_TARGETS = 1 << (u64::BITS - 1);
-    }
-}
-
-impl BaseMeshPipelineKey {
-    pub const PRIMITIVE_TOPOLOGY_MASK_BITS: u64 = 0b111;
-    pub const PRIMITIVE_TOPOLOGY_SHIFT_BITS: u64 =
-        (u64::BITS - 1 - Self::PRIMITIVE_TOPOLOGY_MASK_BITS.count_ones()) as u64;
-
-    pub fn from_primitive_topology(primitive_topology: PrimitiveTopology) -> Self {
-        let primitive_topology_bits = ((primitive_topology as u64)
-            & Self::PRIMITIVE_TOPOLOGY_MASK_BITS)
-            << Self::PRIMITIVE_TOPOLOGY_SHIFT_BITS;
-        Self::from_bits_retain(primitive_topology_bits)
-    }
-
-    pub fn primitive_topology(&self) -> PrimitiveTopology {
-        let primitive_topology_bits = (self.bits() >> Self::PRIMITIVE_TOPOLOGY_SHIFT_BITS)
-            & Self::PRIMITIVE_TOPOLOGY_MASK_BITS;
-        match primitive_topology_bits {
-            x if x == PrimitiveTopology::PointList as u64 => PrimitiveTopology::PointList,
-            x if x == PrimitiveTopology::LineList as u64 => PrimitiveTopology::LineList,
-            x if x == PrimitiveTopology::LineStrip as u64 => PrimitiveTopology::LineStrip,
-            x if x == PrimitiveTopology::TriangleList as u64 => PrimitiveTopology::TriangleList,
-            x if x == PrimitiveTopology::TriangleStrip as u64 => PrimitiveTopology::TriangleStrip,
-            _ => PrimitiveTopology::default(),
-        }
-    }
-}
-
-/// The render world representation of a [`Mesh`].
-#[derive(Debug, Clone)]
-pub struct RenderMesh {
-    /// The number of vertices in the mesh.
-    pub vertex_count: u32,
-
-    /// Morph targets for the mesh, if present.
-    pub morph_targets: Option<TextureView>,
-
-    /// Information about the mesh data buffers, including whether the mesh uses
-    /// indices or not.
-    pub buffer_info: RenderMeshBufferInfo,
-
-    /// Precomputed pipeline key bits for this mesh.
-    pub key_bits: BaseMeshPipelineKey,
-
-    /// A reference to the vertex buffer layout.
-    ///
-    /// Combined with [`RenderMesh::buffer_info`], this specifies the complete
-    /// layout of the buffers associated with this mesh.
-    pub layout: MeshVertexBufferLayoutRef,
-}
-
-impl RenderMesh {
-    /// Returns the primitive topology of this mesh (triangles, triangle strips,
-    /// etc.)
-    #[inline]
-    pub fn primitive_topology(&self) -> PrimitiveTopology {
-        self.key_bits.primitive_topology()
-    }
-}
-
-/// The index/vertex buffer info of a [`RenderMesh`].
-#[derive(Debug, Clone)]
-pub enum RenderMeshBufferInfo {
-    Indexed {
-        count: u32,
-        index_format: IndexFormat,
-    },
-    NonIndexed,
-}
-
-impl RenderAsset for RenderMesh {
-    type SourceAsset = Mesh;
-    type Param = (
-        SRes<RenderAssets<GpuImage>>,
-        SResMut<MeshVertexBufferLayouts>,
-    );
-
-    #[inline]
-    fn asset_usage(mesh: &Self::SourceAsset) -> RenderAssetUsages {
-        mesh.asset_usage
-    }
-
-    fn byte_len(mesh: &Self::SourceAsset) -> Option<usize> {
-        let mut vertex_size = 0;
-        for attribute_data in mesh.attributes.values() {
-            let vertex_format = attribute_data.attribute.format;
-            vertex_size += vertex_format.get_size() as usize;
-        }
-
-        let vertex_count = mesh.count_vertices();
-        let index_bytes = mesh.get_index_buffer_bytes().map(<[_]>::len).unwrap_or(0);
-        Some(vertex_size * vertex_count + index_bytes)
-    }
-
-    /// Converts the extracted mesh into a [`RenderMesh`].
-    fn prepare_asset(
-        mesh: Self::SourceAsset,
-        (images, ref mut mesh_vertex_buffer_layouts): &mut SystemParamItem<Self::Param>,
-    ) -> Result<Self, PrepareAssetError<Self::SourceAsset>> {
-        let morph_targets = match mesh.morph_targets.as_ref() {
-            Some(mt) => {
-                let Some(target_image) = images.get(mt) else {
-                    return Err(PrepareAssetError::RetryNextUpdate(mesh));
-                };
-                Some(target_image.texture_view.clone())
-            }
-            None => None,
-        };
-
-        let buffer_info = match mesh.indices() {
-            Some(indices) => RenderMeshBufferInfo::Indexed {
-                count: indices.len() as u32,
-                index_format: indices.into(),
-            },
-            None => RenderMeshBufferInfo::NonIndexed,
-        };
-
-        let mesh_vertex_buffer_layout =
-            mesh.get_mesh_vertex_buffer_layout(mesh_vertex_buffer_layouts);
-
-        let mut key_bits = BaseMeshPipelineKey::from_primitive_topology(mesh.primitive_topology());
-        key_bits.set(
-            BaseMeshPipelineKey::MORPH_TARGETS,
-            mesh.morph_targets.is_some(),
-        );
-
-        Ok(RenderMesh {
-            vertex_count: mesh.count_vertices() as u32,
-            buffer_info,
-            key_bits,
-            layout: mesh_vertex_buffer_layout,
-            morph_targets,
-        })
-    }
-}
-
-struct MikktspaceGeometryHelper<'a> {
-    indices: Option<&'a Indices>,
-    positions: &'a Vec<[f32; 3]>,
-    normals: &'a Vec<[f32; 3]>,
-    uvs: &'a Vec<[f32; 2]>,
-    tangents: Vec<[f32; 4]>,
-}
-
-impl MikktspaceGeometryHelper<'_> {
-    fn index(&self, face: usize, vert: usize) -> usize {
-        let index_index = face * 3 + vert;
-
-        match self.indices {
-            Some(Indices::U16(indices)) => indices[index_index] as usize,
-            Some(Indices::U32(indices)) => indices[index_index] as usize,
-            None => index_index,
-        }
-    }
-}
-
-impl bevy_mikktspace::Geometry for MikktspaceGeometryHelper<'_> {
-    fn num_faces(&self) -> usize {
-        self.indices
-            .map(Indices::len)
-            .unwrap_or_else(|| self.positions.len())
-            / 3
-    }
-
-    fn num_vertices_of_face(&self, _: usize) -> usize {
-        3
-    }
-
-    fn position(&self, face: usize, vert: usize) -> [f32; 3] {
-        self.positions[self.index(face, vert)]
-    }
-
-    fn normal(&self, face: usize, vert: usize) -> [f32; 3] {
-        self.normals[self.index(face, vert)]
-    }
-
-    fn tex_coord(&self, face: usize, vert: usize) -> [f32; 2] {
-        self.uvs[self.index(face, vert)]
-    }
-
-    fn set_tangent_encoded(&mut self, tangent: [f32; 4], face: usize, vert: usize) {
-        let idx = self.index(face, vert);
-        self.tangents[idx] = tangent;
-    }
-}
-
-#[derive(Error, Debug)]
-/// Failed to generate tangents for the mesh.
-pub enum GenerateTangentsError {
-    #[error("cannot generate tangents for {0:?}")]
-    UnsupportedTopology(PrimitiveTopology),
-    #[error("missing indices")]
-    MissingIndices,
-    #[error("missing vertex attributes '{0}'")]
-    MissingVertexAttribute(&'static str),
-    #[error("the '{0}' vertex attribute should have {1:?} format")]
-    InvalidVertexAttributeFormat(&'static str, VertexFormat),
-    #[error("mesh not suitable for tangent generation")]
-    MikktspaceError,
-}
-
-fn generate_tangents_for_mesh(mesh: &Mesh) -> Result<Vec<[f32; 4]>, GenerateTangentsError> {
-    match mesh.primitive_topology() {
-        PrimitiveTopology::TriangleList => {}
-        other => return Err(GenerateTangentsError::UnsupportedTopology(other)),
-    };
-
-    let positions = mesh.attribute(Mesh::ATTRIBUTE_POSITION).ok_or(
-        GenerateTangentsError::MissingVertexAttribute(Mesh::ATTRIBUTE_POSITION.name),
-    )?;
-    let VertexAttributeValues::Float32x3(positions) = positions else {
-        return Err(GenerateTangentsError::InvalidVertexAttributeFormat(
-            Mesh::ATTRIBUTE_POSITION.name,
-            VertexFormat::Float32x3,
-        ));
-    };
-    let normals = mesh.attribute(Mesh::ATTRIBUTE_NORMAL).ok_or(
-        GenerateTangentsError::MissingVertexAttribute(Mesh::ATTRIBUTE_NORMAL.name),
-    )?;
-    let VertexAttributeValues::Float32x3(normals) = normals else {
-        return Err(GenerateTangentsError::InvalidVertexAttributeFormat(
-            Mesh::ATTRIBUTE_NORMAL.name,
-            VertexFormat::Float32x3,
-        ));
-    };
-    let uvs = mesh.attribute(Mesh::ATTRIBUTE_UV_0).ok_or(
-        GenerateTangentsError::MissingVertexAttribute(Mesh::ATTRIBUTE_UV_0.name),
-    )?;
-    let VertexAttributeValues::Float32x2(uvs) = uvs else {
-        return Err(GenerateTangentsError::InvalidVertexAttributeFormat(
-            Mesh::ATTRIBUTE_UV_0.name,
-            VertexFormat::Float32x2,
-        ));
-    };
-
-    let len = positions.len();
-    let tangents = vec![[0., 0., 0., 0.]; len];
-    let mut mikktspace_mesh = MikktspaceGeometryHelper {
-        indices: mesh.indices(),
-        positions,
-        normals,
-        uvs,
-        tangents,
-    };
-    let success = bevy_mikktspace::generate_tangents(&mut mikktspace_mesh);
-    if !success {
-        return Err(GenerateTangentsError::MikktspaceError);
-    }
-
-    // mikktspace seems to assume left-handedness so we can flip the sign to correct for this
-    for tangent in &mut mikktspace_mesh.tangents {
-        tangent[3] = -tangent[3];
-    }
-
-    Ok(mikktspace_mesh.tangents)
-}
-
-/// Correctly scales and renormalizes an already normalized `normal` by the scale determined by its reciprocal `scale_recip`
-fn scale_normal(normal: Vec3, scale_recip: Vec3) -> Vec3 {
-    // This is basically just `normal * scale_recip` but with the added rule that `0. * anything == 0.`
-    // This is necessary because components of `scale_recip` may be infinities, which do not multiply to zero
-    let n = Vec3::select(normal.cmpeq(Vec3::ZERO), Vec3::ZERO, normal * scale_recip);
-
-    // If n is finite, no component of `scale_recip` was infinite or the normal was perpendicular to the scale
-    // else the scale had at least one zero-component and the normal needs to point along the direction of that component
-    if n.is_finite() {
-        n.normalize_or_zero()
-    } else {
-        Vec3::select(n.abs().cmpeq(Vec3::INFINITY), n.signum(), Vec3::ZERO).normalize()
     }
 }
