+++ crates/bevy_render/src/mesh/primitives/extrusion.rs
@@ -3,9 +3,8 @@ use bevy_math::{
     Vec2, Vec3,
 };
 
-use crate::mesh::{Indices, Mesh, VertexAttributeValues};
-
 use super::{MeshBuilder, Meshable};
+use crate::mesh::{Indices, Mesh, VertexAttributeValues};
 
 /// A type representing a segment of the perimeter of an extrudable mesh.
 pub enum PerimeterSegment {
