+++ crates/bevy_render/src/render_resource/mod.rs
@@ -58,6 +58,8 @@ pub use wgpu::{
     VertexStepMode, COPY_BUFFER_ALIGNMENT,
 };
 
+pub use crate::mesh::VertexBufferLayout;
+
 pub mod encase {
     pub use bevy_encase_derive::ShaderType;
     pub use encase::*;
